requirements.txt
pytest
pandas
openai
python-dotenv